small simplification

+++ src/main.rs
@@ -12,12 +12,12 @@ fn main() {
     let c = {
         let v = 0;
         &v as *const _ as usize
-    };
+    }; // usize
     let d = {
         let v = 0;
         &v as *const _ as usize
-    };
-    let e = (c - d) as usize;
+    }; // usize
+    let e = c - d; // (usize - usize = usize)
     println!("{}==0:{}", e, e == 0);
     println!("{}==0:{}", e, e == 0);
 }
